+++ packages/cli/src/lib/bundler/config.ts
@@ -14,8 +14,13 @@
  * limitations under the License.
  */
 
-import { BackendBundlingOptions, BundlingOptions } from './types';
-import { posix as posixPath, resolve as resolvePath } from 'path';
+import {
+  BackendBundlingOptions,
+  BundlingOptions,
+  ModuleFederationOptions,
+} from './types';
+import { posix as posixPath, resolve as resolvePath, dirname } from 'path';
+import chalk from 'chalk';
 import webpack, { ProvidePlugin } from 'webpack';
 
 import { BackstagePackage } from '@backstage/cli-node';
@@ -28,12 +33,13 @@ import { ModuleFederationPlugin } from '@module-federation/enhanced/webpack';
 import { LinkedPackageResolvePlugin } from './LinkedPackageResolvePlugin';
 import ModuleScopePlugin from 'react-dev-utils/ModuleScopePlugin';
 import { RunScriptWebpackPlugin } from 'run-script-webpack-plugin';
+import ReactRefreshPlugin from '@pmmmwh/react-refresh-webpack-plugin';
 import { paths as cliPaths } from '../../lib/paths';
 import fs from 'fs-extra';
 import { getPackages } from '@manypkg/get-packages';
 import { isChildPath } from '@backstage/cli-common';
 import nodeExternals from 'webpack-node-externals';
-import { optimization } from './optimization';
+import { optimization as optimizationConfig } from './optimization';
 import pickBy from 'lodash/pickBy';
 import { readEntryPoints } from '../entryPoints';
 import { runPlain } from '../run';
@@ -44,13 +50,40 @@ import { hasReactDomClient } from './hasReactDomClient';
 
 const BUILD_CACHE_ENV_VAR = 'BACKSTAGE_CLI_EXPERIMENTAL_BUILD_CACHE';
 
-export function resolveBaseUrl(config: Config): URL {
+export function resolveBaseUrl(
+  config: Config,
+  moduleFederation?: ModuleFederationOptions,
+): URL {
   const baseUrl = config.getOptionalString('app.baseUrl');
+
+  const defaultBaseUrl =
+    moduleFederation?.mode === 'remote'
+      ? `http://localhost:${process.env.PORT ?? '3000'}`
+      : 'http://localhost:3000';
+
   try {
-    return new URL(baseUrl ?? '/', 'http://localhost:3000');
+    return new URL(baseUrl ?? '/', defaultBaseUrl);
   } catch (error) {
     throw new Error(`Invalid app.baseUrl, ${error}`);
   }
+}
+
+export function resolveEndpoint(
+  config: Config,
+  moduleFederation?: ModuleFederationOptions,
+): {
+  host: string;
+  port: number;
+} {
+  const url = resolveBaseUrl(config, moduleFederation);
+
+  return {
+    host: config.getOptionalString('app.listen.host') ?? url.hostname,
+    port:
+      config.getOptionalNumber('app.listen.port') ??
+      Number(url.port) ??
+      (url.protocol === 'https:' ? 443 : 80),
+  };
 }
 
 async function readBuildInfo() {
@@ -93,7 +126,13 @@ export async function createConfig(
   paths: BundlingPaths,
   options: BundlingOptions,
 ): Promise<webpack.Configuration> {
-  const { checksEnabled, isDev, frontendConfig, publicSubPath = '' } = options;
+  const {
+    checksEnabled,
+    isDev,
+    frontendConfig,
+    moduleFederation,
+    publicSubPath = '',
+  } = options;
 
   const { plugins, loaders } = transforms(options);
   // Any package that is part of the monorepo but outside the monorepo root dir need
@@ -101,10 +140,27 @@ export async function createConfig(
   const { packages } = await getPackages(cliPaths.targetDir);
   const externalPkgs = packages.filter(p => !isChildPath(paths.root, p.dir));
 
-  const validBaseUrl = resolveBaseUrl(frontendConfig);
+  const validBaseUrl = resolveBaseUrl(frontendConfig, moduleFederation);
   let publicPath = validBaseUrl.pathname.replace(/\/$/, '');
   if (publicSubPath) {
     publicPath = `${publicPath}${publicSubPath}`.replace('//', '/');
+  }
+
+  if (isDev) {
+    const { host, port } = resolveEndpoint(
+      options.frontendConfig,
+      options.moduleFederation,
+    );
+
+    plugins.push(
+      new ReactRefreshPlugin({
+        overlay: {
+          sockProtocol: 'ws',
+          sockHost: host,
+          sockPort: port,
+        },
+      }),
+    );
   }
 
   if (checksEnabled) {
@@ -232,12 +288,53 @@ export async function createConfig(
     require.resolve('react-refresh'),
   ];
 
+  const mode = isDev ? 'development' : 'production';
+  const optimization = optimizationConfig(options);
+
+  if (
+    mode === 'production' &&
+    process.env.EXPERIMENTAL_MODULE_FEDERATION &&
+    process.env.FORCE_REACT_DEVELOPMENT
+  ) {
+    console.log(
+      chalk.yellow(
+        `⚠️  WARNING: Forcing react and react-dom into development mode. This build should not be used in production.`,
+      ),
+    );
+
+    const reactPackageDirs = [
+      `${dirname(require.resolve('react/package.json'))}/`,
+      `${dirname(require.resolve('react-dom/package.json'))}/`,
+    ];
+
+    // Don't define process.env.NODE_ENV with value matching config.mode. If we
+    // don't set this to false, webpack will define the value of
+    // process.env.NODE_ENV for us, and the definition below will be ignored.
+    optimization.nodeEnv = false;
+
+    // Instead, provide a custom definition which always uses "development" if
+    // the module is part of `react` or `react-dom`, and `config.mode` otherwise.
+    plugins.push(
+      new webpack.DefinePlugin({
+        'process.env.NODE_ENV': webpack.DefinePlugin.runtimeValue(
+          ({ module }) => {
+            if (reactPackageDirs.some(val => module.resource.startsWith(val))) {
+              return '"development"';
+            }
+
+            return `"${mode}"`;
+          },
+        ),
+      }),
+    );
+  }
+
   const withCache = yn(process.env[BUILD_CACHE_ENV_VAR], { default: false });
 
   return {
-    mode: isDev ? 'development' : 'production',
+    mode,
     profile: false,
-    optimization: optimization(options),
+    optimization,
     bail: false,
     performance: {
       hints: false, // we check the gzip size instead