+++ packages/cli/templates/default-backend-plugin/src/service/router.ts
@@ -1,12 +1,11 @@
 import { MiddlewareFactory } from '@backstage/backend-defaults/rootHttpRouter';
-import { LoggerService } from '@backstage/backend-plugin-api';
-import { Config } from '@backstage/config';
+import { LoggerService, RootConfigService } from '@backstage/backend-plugin-api';
 import express from 'express';
 import Router from 'express-promise-router';
 
 export interface RouterOptions {
   logger: LoggerService;
-  config: Config;
+  config: RootConfigService;
 }
 
 export async function createRouter(