+++ plugins/scaffolder-backend-module-cookiecutter/src/actions/fetch/cookiecutter.ts
@@ -28,6 +28,7 @@ import {
   fetchContents,
   executeShellCommand,
 } from '@backstage/plugin-scaffolder-node';
+import { examples } from './cookiecutter.examples';
 
 export class CookiecutterRunner {
   private readonly containerRunner?: ContainerRunner;
@@ -156,6 +157,7 @@ export function createFetchCookiecutterAction(options: {
     id: 'fetch:cookiecutter',
     description:
       'Downloads a template from the given URL into the workspace, and runs cookiecutter on it.',
+    examples,
     schema: {
       input: {
         type: 'object',