+++ plugins/devtools/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-devtools
 
+## 0.1.17-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/frontend-plugin-api@0.6.8-next.0
+  - @backstage/core-components@0.14.10-next.0
+  - @backstage/core-compat-api@0.2.8-next.0
+  - @backstage/core-plugin-api@1.9.3
+  - @backstage/errors@1.2.4
+  - @backstage/plugin-devtools-common@0.1.11
+  - @backstage/plugin-permission-react@0.4.24
+
 ## 0.1.16
 
 ### Patch Changes