+++ plugins/example-todo-list-backend/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @internal/plugin-todo-list-backend
 
+## 1.0.30-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.23.4-next.0
+  - @backstage/backend-plugin-api@0.7.1-next.0
+  - @backstage/errors@1.2.4
+  - @backstage/plugin-auth-node@0.4.18-next.0
+
 ## 1.0.29
 
 ### Patch Changes