+++ plugins/org/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-org
 
+## 0.6.28-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/frontend-plugin-api@0.6.8-next.0
+  - @backstage/plugin-catalog-react@1.12.3-next.0
+  - @backstage/core-components@0.14.10-next.0
+  - @backstage/catalog-model@1.5.0
+  - @backstage/core-compat-api@0.2.8-next.0
+  - @backstage/core-plugin-api@1.9.3
+  - @backstage/plugin-catalog-common@1.0.25
+
 ## 0.6.27
 
 ### Patch Changes